Resolve migrate notes and improve dependencies
