+++ larg4/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(Analysis)
-add_subdirectory(Core)
-add_subdirectory(pluginActions)
 add_subdirectory(Services)
+add_subdirectory(pluginActions)
+add_subdirectory(Core)
+add_subdirectory(Analysis)
