+++ macros/CMakeLists.txt
@@ -1,2 +1,2 @@
-file(GLOB mac_files *.mac)
-install_fw( LIST ${mac_files} )
+file(GLOB mac_files CONFIGURE_DEPENDS *.mac)
+install_fw(LIST ${mac_files})
