+++ ups/product_deps
@@ -252,7 +252,7 @@ artg4tk		v10_05_00	-
 larevt		v09_05_01	-
 nug4		v1_12_02	-
 nurandom	v1_08_01	-
-cetmodules	v3_09_03	-	only_for_build
+cetmodules	v3_12_00	-	only_for_build
 end_product_list
 ####################################
